Continued commit on Py 3.13 upgade

+++ rdagent/components/coder/model_coder/benchmark/gt_code/A-DGN.py
@@ -103,7 +103,7 @@ class AntiSymmetricConv(torch.nn.Module):
 
             if self.act is not None:
                 h = self.act(h)
-
+            h = h.float()
             x = x + self.epsilon * h
 
         return x
